third_party/blink/web_tests/platform/generic/http/tests/websocket/connection-throttling-expected.txt

Test whether WebSocket serialises connections correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS receivedText is "PASS"
PASS receivedText is "PASS"
PASS successfullyParsed is true

TEST COMPLETE
